refactoring token generation
