+++ src/main/java/dz/bououza/service/IConfirmationTokenService.java
@@ -1,0 +1,8 @@
+package dz.bououza.service;
+
+import dz.bououza.domain.AppUser;
+import dz.bououza.domain.ConfirmationToken;
+
+public interface IConfirmationTokenService {
+    ConfirmationToken createNewToken(AppUser user);
+}
